+++ tests/Flow/TransportFlowTest.php
@@ -2,41 +2,32 @@
 
 declare(strict_types=1);
 
-namespace Flow\Test;
-
-use function Amp\delay;
+namespace Flow\Test\Flow;
 
 use ArrayObject;
 use Closure;
-use Flow\Driver\AmpDriver;
 use Flow\DriverInterface;
 use Flow\Flow\Flow;
 use Flow\Flow\TransportFlow;
-use PHPUnit\Framework\TestCase;
+use Flow\IpStrategyInterface;
 use Symfony\Component\Messenger\Envelope;
 use Symfony\Component\Messenger\Transport\InMemoryTransport;
 
-class AmpTransportFlowTest extends TestCase
+class TransportFlowTest extends AbstractFlowTest
 {
     /**
      * @dataProvider jobsProvider
      *
      * @param array<Closure> $jobs
      */
-    public function testJobs(array $jobs, DriverInterface $driver, int $resultNumber): void
+    public function testJobs(DriverInterface $driver, IpStrategyInterface $ipStrategy, array $jobs, int $resultNumber): void
     {
         $transport1 = new InMemoryTransport();
         $transport2 = new InMemoryTransport();
-        $flow = array_reduce($jobs, static function ($flow, $job) use ($driver) {
-            $jobFlow = new Flow($job, static function () {}, null, $driver);
-            if ($flow === null) {
-                $flow = $jobFlow;
-            } else {
-                $flow->fn($jobFlow);
-            }
-
-            return $flow;
-        }, null);
+        $flow = array_reduce(
+            array_map(fn ($job) => new Flow($job, static function () {}, $ipStrategy, $driver), $jobs),
+            fn($flow, $flowIt) => $flow ? $flow->fn($flowIt) : $flowIt
+        );
 
         new TransportFlow($flow, $transport1, $transport2, $driver);
 
@@ -61,28 +52,26 @@ class AmpTransportFlowTest extends TestCase
      */
     public function jobsProvider(): array
     {
-        $driver = new AmpDriver();
-
-        return [
+        return $this->matrix(fn (DriverInterface $driver) => [
             'oneJob' => [[static function (ArrayObject $data): void {
                 $data['number'] = 1;
-            }], $driver, 1],
-            'oneAsyncJob' => [[static function (ArrayObject $data): void {
-                delay(1 / 1000);
+            }], 1],
+            'oneAsyncJob' => [[static function (ArrayObject $data) use ($driver): void {
+                $driver->delay(1 / 1000);
                 $data['number'] = 5;
-            }], $driver, 5],
+            }], 5],
             'twoJob' => [[static function (ArrayObject $data): void {
                 $data['number'] += 2;
             }, static function (ArrayObject $data) {
                 $data['number'] *= 3;
-            }], $driver, 6],
-            'twoAsyncJob' => [[static function (ArrayObject $data): void {
-                delay(1 / 1000);
+            }], 6],
+            'twoAsyncJob' => [[static function (ArrayObject $data) use ($driver): void {
+                $driver->delay(1 / 1000);
                 $data['number'] += 5;
-            }, static function (ArrayObject $data): void {
-                delay(1 / 1000);
+            }, static function (ArrayObject $data) use ($driver): void {
+                $driver->delay(1 / 1000);
                 $data['number'] *= 2;
-            }], $driver, 10],
-        ];
+            }], 10],
+        ]);
     }
 }